prototipo/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
requests>=2.31
openai>=1.12
httpx>=0.26
